Extract unwrap_contiguous_iterator into separate file to avoid nvrtc errors
