+++ thrust/thrust/system/cuda/detail/adjacent_difference.h
@@ -51,6 +51,7 @@
 #  include <thrust/system/cuda/detail/par_to_seq.h>
 #  include <thrust/system/cuda/detail/util.h>
 #  include <thrust/type_traits/is_contiguous_iterator.h>
+#  include <thrust/type_traits/unwrap_contiguous_iterator.h>
 
 #  include <cuda/std/type_traits>
