+++ thrust/thrust/type_traits/is_contiguous_iterator.h
@@ -31,13 +31,7 @@
 #elif defined(_CCCL_IMPLICIT_SYSTEM_HEADER_MSVC)
 #  pragma system_header
 #endif // no system header
-#include <thrust/detail/raw_pointer_cast.h>
-#include <thrust/detail/type_traits.h>
-#include <thrust/detail/type_traits/pointer_traits.h>
-
-#include <iterator>
-#include <type_traits>
-#include <utility>
+#include <thrust/detail/type_traits/is_thrust_pointer.h>
 
 #if _CCCL_COMPILER(MSVC, <, 19, 16) // MSVC 2017 version 15.9
 #  include <array>
@@ -212,70 +206,7 @@ struct is_contiguous_iterator_impl
           || is_msvc_contiguous_iterator<Iterator>::value || proclaim_contiguous_iterator<Iterator>::value>
 {};
 
-// Type traits for contiguous iterators:
-template <typename Iterator>
-struct contiguous_iterator_traits
-{
-  static_assert(thrust::is_contiguous_iterator<Iterator>::value,
-                "contiguous_iterator_traits requires a contiguous iterator.");
-
-  using raw_pointer = typename thrust::detail::pointer_traits<decltype(&*std::declval<Iterator>())>::raw_pointer;
-};
 } // namespace detail
-
-//! Converts a contiguous iterator type to its underlying raw pointer type.
-template <typename ContiguousIterator>
-using unwrap_contiguous_iterator_t = typename detail::contiguous_iterator_traits<ContiguousIterator>::raw_pointer;
-
-//! Converts a contiguous iterator to its underlying raw pointer.
-template <typename ContiguousIterator>
-_CCCL_HOST_DEVICE auto unwrap_contiguous_iterator(ContiguousIterator it)
-  -> unwrap_contiguous_iterator_t<ContiguousIterator>
-{
-  static_assert(thrust::is_contiguous_iterator<ContiguousIterator>::value,
-                "unwrap_contiguous_iterator called with non-contiguous iterator.");
-  return thrust::raw_pointer_cast(&*it);
-}
-
-namespace detail
-{
-// Implementation for non-contiguous iterators -- passthrough.
-template <typename Iterator, bool IsContiguous = thrust::is_contiguous_iterator<Iterator>::value>
-struct try_unwrap_contiguous_iterator_impl
-{
-  using type = Iterator;
-
-  static _CCCL_HOST_DEVICE type get(Iterator it)
-  {
-    return it;
-  }
-};
-
-// Implementation for contiguous iterators -- unwraps to raw pointer.
-template <typename Iterator>
-struct try_unwrap_contiguous_iterator_impl<Iterator, true /*is_contiguous*/>
-{
-  using type = unwrap_contiguous_iterator_t<Iterator>;
-
-  static _CCCL_HOST_DEVICE type get(Iterator it)
-  {
-    return unwrap_contiguous_iterator(it);
-  }
-};
-} // namespace detail
-
-//! Takes an iterator type and, if it is contiguous, yields the raw pointer type it represents. Otherwise returns the
-//! iterator type unmodified.
-template <typename Iterator>
-using try_unwrap_contiguous_iterator_t = typename detail::try_unwrap_contiguous_iterator_impl<Iterator>::type;
-
-//! Takes an iterator and, if it is contiguous, unwraps it to the raw pointer it represents. Otherwise returns the
-//! iterator unmodified.
-template <typename Iterator>
-_CCCL_HOST_DEVICE auto try_unwrap_contiguous_iterator(Iterator it) -> try_unwrap_contiguous_iterator_t<Iterator>
-{
-  return detail::try_unwrap_contiguous_iterator_impl<Iterator>::get(it);
-}
 
 /*! \endcond
  */
